Time: 1142 ms (10.53%), Space: 8.1 MB (79.70%) - LeetHub

+++ 1689-partitioning-into-minimum-number-of-deci-binary-numbers/1689-partitioning-into-minimum-number-of-deci-binary-numbers.c
@@ -4,7 +4,7 @@ int minPartitions(char* n) {
     // 문자열로 이루어진 n을 분리해서 nums에 저장 (이때 int로 형변환)
     int nums[100000];
     for (int i=0; i<strlen(n); i++) {
-        nums[i] = (int)(n[i] - '0');
+        nums[i] = n[i] - '0';
     }
 
     // 그 중 최대값 찾아서 return
